pyproject.toml
[project]
name = "basketball-box-score-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "requests>=2.32.5",
    "opencv-python",
    "numpy",
    "torch",
    "torchvision",
    "python-dotenv",
    "inference>=0.62.2",
    "sports",
]

[project.scripts]
hoopstats = "hoopstats.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/hoopstats"]

[tool.hatch.dependencies]
python-dotenv = "*"

[tool.uv.sources]
sports = { git = "https://github.com/roboflow/sports.git", rev = "feat/basketball" }
